# app/Http/Requests/UpdateResponseItemRequest.php
<?php
/**
 * Created by PhpStorm.
 * User: kirya
 * Date: 18.09.16
 * Time: 12:50
 */

namespace App\Http\Requests;

use App\Demand\ResponseItem;
use App\Services\ResponseItemService;
use Illuminate\Http\Request;
use Swagger\Annotations as SWG;

/**
 * @SWG\Definition(
 *      definition="UpdateResponseItemRequest",
 *      required={"price"},
 *      @SWG\Property(property="price", type="number")
 * )
 *
 * Class UpdateResponseItemRequest
 * @package App\Http\Requests
 */
class UpdateResponseItemRequest extends ApiRequest
{
    /**
     * @var ResponseItem
     */
    private $item;

    /**
     * @var ResponseItemService
     */
    private $responseItemService;

    public function __construct(array $query = array(), array $request = array(), array $attributes = array(), array $cookies = array(), array $files = array(), array $server = array(), $content = null)
    {
        $this->responseItemService = new ResponseItemService();
        parent::__construct($query, $request, $attributes, $cookies, $files, $server, $content); // TODO: Change the autogenerated stub
    }


    /**
     * Determine if the user is authorized to make this request.
     *
     * @return bool
     */
    public function authorize()
    {
        return $this->user()->can('update', $this->getResponseItem());
    }

    /**
     * Get the validation rules that apply to the request.
     *
     * @return array
     */
    public function rules()
    {
        return [
            'price' => 'numeric|required'
        ];
    }

    /**
     * @return ResponseItem
     */
    public function getResponseItem()
    {
        if (!isset($this->item)) {
            $this->item = $this->responseItemService->findItem((int)$this->route('id'));
        }
        return $this->item;
    }


}