# public/doc/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Supp",
        "description": "An API that uses supp service",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/demandItems/{id}": {
            "patch": {
                "tags": [
                    "demandItem"
                ],
                "summary": "Update demandItem",
                "description": "",
                "operationId": "updateDemandItem",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "DemandItem id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "DemandItem",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/UpdateDemandItemRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "successful operation"
                    },
                    "404": {
                        "$ref": "#/responses/NotFoundResponse"
                    },
                    "401": {
                        "$ref": "#/responses/NotAuthResponse"
                    },
                    "default": {
                        "$ref": "#/responses/DefaultErrorResponse"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/demands": {
            "get": {
                "tags": [
                    "demand"
                ],
                "summary": "Get my demands by status",
                "description": "",
                "operationId": "getDemands",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "filtered Status [or array of status]",
                        "type": "string",
                        "enum": [
                            "active",
                            "archived"
                        ]
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/DemandModelWithResponses"
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/responses/NotFoundResponse"
                    },
                    "401": {
                        "$ref": "#/responses/NotAuthResponse"
                    },
                    "default": {
                        "$ref": "#/responses/DefaultErrorResponse"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "demand"
                ],
                "summary": "Create demand",
                "description": "",
                "operationId": "createDemands",
                "parameters": [
                    {
                        "name": "Demand",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/CreateDemandRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "headers": {
                            "location": {
                                "description": "/demands/1",
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/responses/NotFoundResponse"
                    },
                    "401": {
                        "$ref": "#/responses/NotAuthResponse"
                    },
                    "default": {
                        "$ref": "#/responses/DefaultErrorResponse"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/demands/{id}": {
            "patch": {
                "tags": [
                    "demand"
                ],
                "summary": "Update demand",
                "description": "",
                "operationId": "updateDemands",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "Demand",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/UpdateDemandRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "headers": {
                            "location": {
                                "description": "/demands/1",
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/responses/NotFoundResponse"
                    },
                    "401": {
                        "$ref": "#/responses/NotAuthResponse"
                    },
                    "default": {
                        "$ref": "#/responses/DefaultErrorResponse"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
        "/responseItems/{id}": {
            "delete": {
                "tags": [
                    "responseItem"
                ],
                "summary": "Delete responseItem",
                "description": "",
                "operationId": "deleteResponseItem",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "202": {
                        "description": "successful operation"
                    },
                    "404": {
                        "$ref": "#/responses/NotFoundResponse"
                    },
                    "401": {
                        "$ref": "#/responses/NotAuthResponse"
                    },
                    "default": {
                        "$ref": "#/responses/DefaultErrorResponse"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "responseItem"
                ],
                "summary": "Update ResponseItem",
                "description": "",
                "operationId": "updateResponseItem",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "type": "integer"
                    },
                    {
                        "name": "ResponseItem",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/UpdateResponseItemRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "successful operation"
                    },
                    "404": {
                        "$ref": "#/responses/NotFoundResponse"
                    },
                    "401": {
                        "$ref": "#/responses/NotAuthResponse"
                    },
                    "default": {
                        "$ref": "#/responses/DefaultErrorResponse"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
        "ErrorModel": {
            "required": [
                "status_code",
                "message"
            ],
            "properties": {
                "status_code": {
                    "type": "integer",
                    "format": "int32"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "CreateDemandRequest": {
            "required": [
                "demandItems"
            ],
            "properties": {
                "title": {
                    "description": "title demand",
                    "type": "string"
                },
                "address": {
                    "description": "address demand",
                    "type": "string"
                },
                "desc": {
                    "description": "description of demand",
                    "type": "string"
                },
                "delivery_date": {
                    "description": "date delivery of demand(dd.mm.YYYY)",
                    "type": "string"
                },
                "addition_emails": {
                    "description": "addition emails",
                    "type": "array",
                    "items": {
                        "description": "email",
                        "type": "string"
                    }
                },
                "regions": {
                    "description": "regions",
                    "type": "array",
                    "items": {
                        "description": "region",
                        "type": "string"
                    }
                },
                "spheres": {
                    "description": "spheres",
                    "type": "array",
                    "items": {
                        "description": "sphere",
                        "type": "string"
                    }
                },
                "demandItems": {
                    "description": "demandItems",
                    "type": "array",
                    "items": {
                        "description": "demandItem",
                        "required": [
                            "title",
                            "count",
                            "quantity_id"
                        ],
                        "properties": {
                            "title": {
                                "type": "string"
                            },
                            "count": {
                                "type": "number"
                            },
                            "quantity_id": {
                                "type": "integer"
                            }
                        },
                        "type": "object"
                    }
                }
            }
        },
...
        "UpdateDemandItemRequest": {
            "properties": {
                "response_item_id": {
                    "description": "select or unselect(null)",
                    "type": "integer"
                }
            }
        },
        "UpdateDemandRequest": {
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string",
                    "enum": [
                        "active",
                        "archived"
                    ]
                }
            }
        },
        "UpdateResponseItemRequest": {
            "required": [
                "price"
            ],
            "properties": {
                "price": {
                    "type": "number"
                }
            }
        },
...
        "CompanyModel": {
            "properties": {
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "DemandItemModel": {
            "properties": {
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "demand_id": {
                    "type": "integer"
                },
                "quantityTitle": {
                    "description": "string or null",
                    "type": "string"
                },
                "count": {
                    "type": "number"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "DemandModelWithResponses": {
            "properties": {
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "company": {
                    "type": "object",
                    "$ref": "#/definitions/CompanyModel"
                },
                "demandItems": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/DemandItemModel"
                    }
                },
                "responses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ResponseModel"
                    }
                }
            }
        },
        "InvoiceModel": {
            "properties": {
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "response_id": {
                    "type": "integer"
                },
                "file": {
                    "type": "string"
                }
            }
        },
        "ResponseItemModel": {
            "properties": {
                "id": {
                    "type": "integer"
                },
                "response_id": {
                    "type": "integer"
                },
                "price": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "selected": {
                    "type": "boolean"
                },
                "invoice_id": {
                    "type": "integer"
                },
                "demand_item_id": {
                    "type": "integer"
                }
            }
        },
        "ResponseModel": {
            "properties": {
                "id": {
                    "description": "/**",
                    "type": "integer"
                },
                "demand_id": {
                    "description": "/**",
                    "type": "integer"
                },
                "delivery_type": {
                    "description": "/**",
                    "type": "string"
                },
                "status": {
                    "description": "/**",
                    "type": "string"
                },
                "company": {
                    "description": "/**",
                    "type": "object",
                    "$ref": "#/definitions/CompanyModel"
                },
                "invoices": {
                    "description": "/**",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/InvoiceModel"
                    }
                },
                "responseItems": {
                    "description": "/**",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ResponseItemModel"
                    }
                }
            }
        },
...
    },
    "responses": {
        "NotAuthResponse": {
            "description": "Not Auth",
            "schema": {
                "$ref": "#/definitions/ErrorModel"
            }
        },
        "NotFoundResponse": {
            "description": "Not Found",
            "schema": {
                "$ref": "#/definitions/ErrorModel"
            }
        },
        "DefaultErrorResponse": {
            "description": "Default Error",
            "schema": {
                "$ref": "#/definitions/ErrorModel"
            }
        }
    },
    "securityDefinitions": {
        "token": {
            "type": "apiKey",
            "name": "token",
            "in": "query"
        }
    }
}